Remove old list and move messages under GiftCardList
