--- src/giftCards/GiftCardsList/GiftCardsListTable/GiftCardsListTableHeader/index.tsx
@@ -1,2 +1,0 @@
-export * from "./GiftCardsListTableHeader";
-export { default } from "./GiftCardsListTableHeader";
