--- src/giftCards/GiftCardsList/GiftCardsListTable/index.tsx
@@ -1,2 +1,0 @@
-export * from "./GiftCardsListTable";
-export { default } from "./GiftCardsListTable";
